refactor(session): Rename and apply vaccel type changes
Update implementation to apply vaccel naming and type changes for
session object:
- Update operations to use renamed `vaccel_session_*()` functions
- Change session id type to vaccel_id_t/VaccelId where needed
- Rename session client functions with the `vaccel_rpc_client` prefix

PR: https://github.com/nubificus/vaccel-rust/pull/5
Approved-by: Anastassios Nanos <ananos@nubificus.co.uk>

+++ vaccel-rpc-client/src/asynchronous/client.rs
@@ -15,7 +15,7 @@ use vaccel_rpc_proto::asynchronous::agent_ttrpc::RpcAgentClient;
 #[repr(C)]
 pub struct VaccelRpcClient {
     pub ttrpc_client: RpcAgentClient,
-    pub timers: Arc<DashMap<u32, ProfRegions>>,
+    pub timers: Arc<DashMap<i64, ProfRegions>>,
     pub runtime: Arc<Runtime>,
 }
