+++ vaccel-rpc-client/src/sync/client.rs
@@ -14,7 +14,7 @@ use vaccel_rpc_proto::sync::agent_ttrpc::RpcAgentClient;
 #[repr(C)]
 pub struct VaccelRpcClient {
     pub ttrpc_client: RpcAgentClient,
-    pub timers: Arc<DashMap<u32, ProfRegions>>,
+    pub timers: Arc<DashMap<i64, ProfRegions>>,
 }
 
 impl VaccelRpcClient {
